Move AssertExpectedRange to main Fixed/Fixed128 files
